requirements.txt
simpy>=4.0.1
pyyaml>=6.0
numpy>=1.24.0
pandas>=2.0.0
pytest>=7.4.0
pytest-cov>=4.1.0
